requirements.txt
langchain
openai
pandas
pinecone-client
python-dotenv
streamlit
tiktoken